+++ src/components/generation/passage/PassageContent/PassageContentMain.vue
