+++ src/components/generation/passage/PassageContent/PassageSummary.vue
