+++ src/components/generation/passage/PassageMain.vue
@@ -15,8 +15,8 @@
     </div>
 </template>
 <script setup>
-import CreatePassageMain from '@/components/generation/passage/CreatePassageMain.vue';
-import PaymentUsage from '@/components/generation/passage/PaymentUsage.vue';
+import CreatePassageMain from '@/components/generation/passage/PassageMain/CreatePassageMain.vue';
+import PaymentUsage from '@/components/generation/passage/PassageMain/PaymentUsage.vue';
 import BaseButton from '../common/button/BaseButton.vue';
 </script>
 <style scoped>
