+++ src/components/generation/passage/PassageMain/CreatePassageMain.vue
